datalens-main/requirements.txt
pandas==2.1.0
numpy==1.24.3
streamlit==1.26.0
plotly==5.16.1
scikit-learn==1.3.0
langchain==0.0.27
google-cloud-aiplatform==1.34.0
vertexai==0.0.1
python-dotenv==1.0.0
matplotlib==3.7.2
seaborn==0.12.2
jupyter==1.0.0
google-cloud-storage==2.10.0